Extract golden apple and out of bounds code to shared lib
