+++ src/bin/minibit/subservers/boxing.rs
@@ -45,7 +45,6 @@ pub fn main(path: PathBuf) {
             Update,
             (
                 init_clients.after(minibit_lib::duels::map::init_clients::<DefaultDuelsConfig>),
-                handle_oob_clients,
                 end_game.after(minibit_lib::duels::map::end_game::<DefaultDuelsConfig>),
             ),
         )
@@ -161,20 +160,6 @@ fn handle_combat_events(
             end_game.send(EndGameEvent {
                 game_id: victim.gamestate.game_id.unwrap(),
                 loser: victim.gamestate.team,
-            });
-        }
-    }
-}
-
-fn handle_oob_clients(
-    positions: Query<(&mut Position, &PlayerGameState), With<Client>>,
-    mut end_game: EventWriter<EndGameEvent>,
-) {
-    for (pos, gamestate) in positions.iter() {
-        if pos.0.y < 0.0 && let Some(game_id) = gamestate.game_id {
-            end_game.send(EndGameEvent {
-                game_id,
-                loser: gamestate.team,
             });
         }
     }
