+++ src/bin/minibit/subservers/bridge.rs
@@ -37,10 +37,7 @@ use valence::entity::living::Health;
 use valence::entity::Velocity;
 use valence::entity::{EntityId, EntityStatuses};
 use valence::equipment::EquipmentInventorySync;
-use valence::event_loop::PacketEvent;
-use valence::interact_item::InteractItemEvent;
 use valence::inventory::HeldItem;
-use valence::inventory::PlayerAction;
 use valence::math::IVec3;
 use valence::math::Vec3Swizzles;
 use valence::nbt::compound;
@@ -48,12 +45,13 @@ use valence::nbt::List;
 use valence::prelude::*;
 use valence::protocol::packets::play::DamageTiltS2c;
 use valence::protocol::packets::play::ExperienceBarUpdateS2c;
-use valence::protocol::packets::play::PlayerActionC2s;
 use valence::protocol::sound::SoundCategory;
 use valence::protocol::Sound;
 use valence::protocol::VarInt;
 use valence::protocol::WritePacket;
 use minibit_lib::death::{DeathEvent, DeathPlugin, DeathSet};
+use minibit_lib::duels::oob::{OobMode, OobPlugin};
+use minibit_lib::food::golden_apple::GoldenApplePlugin;
 
 #[derive(Event)]
 struct ScoreEvent (Entity);
@@ -92,9 +90,6 @@ impl DuelsConfig for BridgeConfig {
     }
 }
 
-#[derive(Component)]
-struct EatingStartTick(pub i64);
-
 pub fn main(path: PathBuf) {
     App::new()
         .add_plugins(DuelsPlugin::<BridgeConfig> {
@@ -109,6 +104,7 @@ pub fn main(path: PathBuf) {
             DisableDropPlugin,
             ProjectilePlugin,
             DeathPlugin,
+            GoldenApplePlugin,
             DiggingPlugin {
                 whitelist: vec![
                     BlockKind::BlueTerracotta,
@@ -124,6 +120,10 @@ pub fn main(path: PathBuf) {
                 max_z: 60,
                 min_z: -60,
             },
+            OobPlugin {
+                mode: OobMode::DeathEvent,
+                bounds_y: 75.0..,
+            },
         ))
         .add_event::<ScoreEvent>()
         .add_event::<MessageEvent>()
@@ -138,13 +138,9 @@ pub fn main(path: PathBuf) {
                 end_game,
                 check_goals,
                 update_bow_cooldown,
-                set_use_tick,
-                eat_gapple,
-                cancel_gapple,
                 handle_collision_events,
                 handle_death.after(DeathSet),
                 handle_score.after(check_goals).before(handle_death),
-                handle_oob_clients,
                 game_broadcast,
             ),
         )
@@ -161,7 +157,7 @@ fn setup(mut commands: Commands, server_config: Res<BridgeConfig>) {
 
 fn init_clients(clients: Query<Entity, Added<Client>>, mut commands: Commands) {
     for entity in clients.iter() {
-        commands.entity(entity).insert((EquipmentInventorySync, EatingStartTick(i64::MAX), BowStatus::default()));
+        commands.entity(entity).insert((EquipmentInventorySync, BowStatus::default()));
     }
 }
 
@@ -363,69 +359,6 @@ fn update_bow_cooldown(
     }
 }
 
-fn set_use_tick(
-    mut clients: Query<(&Inventory, &HeldItem, &mut EatingStartTick), With<Client>>,
-    mut events: EventReader<InteractItemEvent>,
-    server: Res<Server>,
-) {
-    for event in events.read() {
-        if let Ok((inv, held_item, mut eat_tick)) = clients.get_mut(event.client)
-            && event.hand == Hand::Main
-            && inv.slot(held_item.slot()).item == ItemKind::GoldenApple
-        {
-            eat_tick.0 = server.current_tick();
-        }
-    }
-}
-
-fn eat_gapple(
-    mut clients: Query<
-        (
-            &mut Client,
-            &mut Health,
-            &mut Absorption,
-            &mut Inventory,
-            &HeldItem,
-            &mut EatingStartTick,
-        ),
-        With<Client>,
-    >,
-    server: Res<Server>,
-) {
-    for (mut client, mut health, mut absorption, mut inv, held_item, mut eat_tick) in
-        clients.iter_mut()
-    {
-        let slot = held_item.slot();
-        if inv.slot(slot).item != ItemKind::GoldenApple {
-            eat_tick.0 = i64::MAX;
-            continue;
-        }
-        if server.current_tick() - eat_tick.0 > 32 {
-            eat_tick.0 = i64::MAX;
-            client.trigger_status(EntityStatus::ConsumeItem);
-            health.0 = 20.0;
-            absorption.0 = 4.0;
-            let count = inv.slot(slot).count;
-            inv.set_slot_amount(held_item.slot(), count - 1);
-        }
-    }
-}
-
-fn cancel_gapple(
-    mut clients: Query<(&Inventory, &HeldItem, &mut EatingStartTick), With<Client>>,
-    mut packets: EventReader<PacketEvent>,
-) {
-    for packet in packets.read() {
-        if let Some(pkt) = packet.decode::<PlayerActionC2s>()
-            && pkt.action == PlayerAction::ReleaseUseItem
-            && let Ok((inv, held_item, mut eat_tick)) = clients.get_mut(packet.client)
-            && inv.slot(held_item.slot()).item == ItemKind::GoldenApple
-        {
-            eat_tick.0 = i64::MAX;
-        }
-    }
-}
-
 #[derive(QueryData)]
 #[query_data(mutable)]
 struct CombatQuery {
@@ -551,17 +484,6 @@ fn handle_collision_events(
                 1.0,
                 1.0,
             );
-        }
-    }
-}
-
-fn handle_oob_clients(
-    positions: Query<(Entity, &Position, &PlayerGameState), With<Client>>,
-    mut deaths: EventWriter<DeathEvent>,
-) {
-    for (entity, pos, gamestate) in positions.iter() {
-        if pos.0.y < 75.0 && gamestate.game_id.is_some() {
-            deaths.send(DeathEvent(entity, true));
         }
     }
 }
