+++ src/bin/minibit/subservers/classic.rs
@@ -32,13 +32,18 @@ use valence::protocol::sound::SoundCategory;
 use valence::protocol::Sound;
 use valence::protocol::VarInt;
 use valence::protocol::WritePacket;
+use minibit_lib::duels::oob::{OobMode, OobPlugin};
 
 pub fn main(path: PathBuf) {
     App::new()
         .add_plugins(DuelsPlugin::<DefaultDuelsConfig> { path, default_gamemode: GameMode::Adventure, copy_map: false, phantom: PhantomData })
         .add_plugins(DefaultPlugins)
+        .add_plugins(OobPlugin {
+            mode: OobMode::GameEndEvent,
+            bounds_y: 0.0..,
+        })
         .add_systems(EventLoopUpdate, handle_combat_events)
-        .add_systems(Update, (start_game, end_game, handle_oob_clients))
+        .add_systems(Update, (start_game, end_game))
         .run();
 }
 
@@ -177,19 +182,5 @@ fn handle_combat_events(
             entity_id: VarInt(victim.id.get()),
             yaw: 0.0,
         });
-    }
-}
-
-fn handle_oob_clients(
-    positions: Query<(&Position, &PlayerGameState), With<Client>>,
-    mut end_game: EventWriter<EndGameEvent>,
-) {
-    for (pos, gamestate) in positions.iter() {
-        if pos.0.y < 0.0 && let Some(game_id) = gamestate.game_id {
-            end_game.send(EndGameEvent {
-                game_id,
-                loser: gamestate.team,
-            });
-        }
     }
 }
