+++ src/lib/food/mod.rs
@@ -1,0 +1,1 @@
+pub mod golden_apple;
