+++ src/lib/mod.rs
@@ -26,3 +26,4 @@ pub mod player;
 pub mod projectiles;
 pub mod scopes;
 pub mod world;
+pub mod food;
